Remove generated 'WorkspacesUpdateTags' and replace with a custom version.
Acceptance test output:

$ make testacc TEST=./aws/ TESTARGS='-run=TestAccAwsWorkspacesDirectory/basic'
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go test ./aws/ -v -count 1 -parallel 20 -run=TestAccAwsWorkspacesDirectory/basic -timeout 120m
=== RUN   TestAccAwsWorkspacesDirectory
=== RUN   TestAccAwsWorkspacesDirectory/basic
--- PASS: TestAccAwsWorkspacesDirectory (661.39s)
    --- PASS: TestAccAwsWorkspacesDirectory/basic (661.39s)
PASS
ok  	github.com/terraform-providers/terraform-provider-aws/aws	661.429s
