+++ aws/resource_aws_workspaces_directory.go
@@ -118,8 +118,6 @@ func resourceAwsWorkspacesDirectoryCreate(d *schema.ResourceData, meta interface
 	}
 	log.Printf("[DEBUG] Workspaces directory %q is registered", d.Id())
 
-	d.Partial(true)
-
 	log.Printf("[DEBUG] Modifying workspaces directory %q self-service permissions...", d.Id())
 	if v, ok := d.GetOk("self_service_permissions"); ok {
 		_, err := conn.ModifySelfservicePermissions(&workspaces.ModifySelfservicePermissionsInput{
@@ -129,11 +127,8 @@ func resourceAwsWorkspacesDirectoryCreate(d *schema.ResourceData, meta interface
 		if err != nil {
 			return fmt.Errorf("error setting self service permissions: %s", err)
 		}
-		d.SetPartial("self_service_permission")
 	}
 	log.Printf("[DEBUG] Workspaces directory %q self-service permissions are set", d.Id())
-
-	d.Partial(false)
 
 	return resourceAwsWorkspacesDirectoryRead(d, meta)
 }
@@ -173,8 +168,6 @@ func resourceAwsWorkspacesDirectoryRead(d *schema.ResourceData, meta interface{}
 func resourceAwsWorkspacesDirectoryUpdate(d *schema.ResourceData, meta interface{}) error {
 	conn := meta.(*AWSClient).workspacesconn
 
-	d.Partial(true)
-
 	if d.HasChange("self_service_permissions") {
 		log.Printf("[DEBUG] Modifying workspaces directory %q self-service permissions...", d.Id())
 		permissions := d.Get("self_service_permissions").([]interface{})
@@ -187,20 +180,14 @@ func resourceAwsWorkspacesDirectoryUpdate(d *schema.ResourceData, meta interface
 			return fmt.Errorf("error updating self service permissions: %s", err)
 		}
 		log.Printf("[DEBUG] Workspaces directory %q self-service permissions are set", d.Id())
-		d.SetPartial("self_service_permission")
 	}
 
 	if d.HasChange("tags") {
-		log.Printf("[DEBUG] Modifying workspaces directory %q tags...", d.Id())
 		o, n := d.GetChange("tags")
-		if err := WorkspacesUpdateTags(conn, d.Id(), o, n); err != nil {
+		if err := keyvaluetags.WorkspacesUpdateTags(conn, d.Id(), o, n); err != nil {
 			return fmt.Errorf("error updating tags: %s", err)
 		}
-		log.Printf("[DEBUG] Workspaces directory %q tags are modified", d.Id())
-		d.SetPartial("tags")
 	}
-
-	d.Partial(false)
 
 	return resourceAwsWorkspacesDirectoryRead(d, meta)
 }
